requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
Brotli==1.0.9
cffi==1.14.6
cssselect2==0.4.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
fonttools==4.27.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
Markdown==3.3.4
numpy==1.21.2
openpyxl==3.0.9
Pillow==8.3.2
platformdirs==2.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
pydyf==0.1.1
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
python-decouple==3.5
pytz==2021.1
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
values==2020.12.3
virtualenv==20.6.0
weasyprint==53.3
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
zopfli==0.1.8
